Added screenshot
Here is the executed HTML code. It's coming together, but I noticed a few unintentional errors. I didn't think I would need the quotation marks to define the text, but just in case I put them in my body and title. I will remove those and continue to add to the text.